TritonDataCenter/zfs_snapshot_tar#6 Want 64-bit compatibility (#7)
joyent/zfs_snapshot_tar#6 Want 64-bit compatibility
Reviewed by: Dan McDonald <danmcd@mnx.io>

+++ deps/illumos/run_command.c
@@ -11,6 +11,7 @@
 
 /*
  * Copyright 2015 Joyent, Inc.
+ * Copyright 2024 MNX Cloud, Inc.
  */
 
 #include <stdlib.h>
@@ -201,7 +202,7 @@ static int
 cb_data(const uint8_t *buf, size_t sz, void *arg0, void *arg1)
 {
 	cmd_t *cmd = arg0;
-	int fd = (int)arg1;
+	int fd = (int)(uintptr_t)arg1;
 	unsigned int i;
 
 	if (cmd->cmd_cancel) {
@@ -258,7 +259,7 @@ static void
 cb_eof(void *arg0, void *arg1)
 {
 	cmd_t *cmd = arg0;
-	int fd = (int)arg1;
+	int fd = (int)(uintptr_t)arg1;
 
 	if (cmd->cmd_cancel) {
 		return;
@@ -274,7 +275,7 @@ static void
 cb_error(int e, void *arg0, void *arg1)
 {
 	cmd_t *cmd = arg0;
-	int fd = (int)arg1;
+	int fd = (int)(uintptr_t)arg1;
 
 	if (cmd->cmd_cancel) {
 		return;
